[2235][AD] Avoid invalid notifications for unexisting resources
On Team for Capella, with shared resource and a project without the
Capella nature, Kitalpha tries to unreference the ad Viewpoint, as it
detects that the Capella nature is not there.

But here is no afm resource, and the attempt to load an unexisting
shared (cdo://) afm resource leads to an IllegalStateException which is
properly catched line 308.

Then the resource is removed whereas it has never been added to the
resource set. This leads to invalid ADD/REMOVE notifications throwing
IllegalStateException when Sirius / Team for Capella tries to analyse
them.

This commits checks with the URIConverter of the resource set if the
resource exists before trying to load it.

It also prevents the unReference method to create the .afm resource if
it does not exist.

Bug: 2235
Change-Id: I51526a033f108bc10f82fb15682adb08b7e51e1e
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>