[Rebase] v1.2.x to v1.2.2_for_capella_v1.2.1
[2273][Detachment] Cleaning unknown references does not detect/delete
Sirius Tables.

- Add Test
- Fix viewpoint core tests
- cherry-pick from 1.3.x to v1.2.x branch

bug: 2273
Change-Id: Ie1582e9cb498f9250aa60d6460f09755205ee42c
Signed-off-by: Faycal Abka <abka.faycal@gmail.com>

Change-Id: I4b8c69df04206854fedb7251e3047778536339d5
Signed-off-by: Faycal Abka <abka.faycal@gmail.com>

[2235][AD] Avoid invalid notifications for unexisting resources

On Team for Capella, with shared resource and a project without the
Capella nature, Kitalpha tries to unreference the ad Viewpoint, as it
detects that the Capella nature is not there.

But here is no afm resource, and the attempt to load an unexisting
shared (cdo://) afm resource leads to an IllegalStateException which is
properly catched line 308.

Then the resource is removed whereas it has never been added to the
resource set. This leads to invalid ADD/REMOVE notifications throwing
IllegalStateException when Sirius / Team for Capella tries to analyse
them.

This commits checks with the URIConverter of the resource set if the
resource exists before trying to load it.

It also prevents the unReference method to create the .afm resource if
it does not exist.

Bug: 2235
Change-Id: I51526a033f108bc10f82fb15682adb08b7e51e1e
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>

[2285] CCE when cast to Metadata

bug: 2285

Change-Id: Ifbd520ae9ccb5e30170d5abbf0861983d1fb1749
Signed-off-by: Thomas Guiu <thomas.guiu@soyatec.com>

[2356][GenDoc] Add mechanism to enhance the ergonomy of search index

bug: 2356
Change-Id: Id96b3c981c3c778a5ea59e8af54710edfaeb8ce3
Signed-off-by: Faycal Abka <abka.faycal@gmail.com>

[1824] Add ability to select sirius representations to export

- Commit reworked from https://git.polarsys.org/r/#/c/5331/
- The contribution cannot be merged due to non signing of CLA by the
contributor
- Author: Yann Mortier <yann.mortier@obeo.fr>

bug: 1824
Change-Id: Iae76470e559cbfd29a7ceb69390be33d35ca2b80
Signed-off-by: Faycal Abka <abka.faycal@gmail.com>

[2356][GenDoc] Add mechanism to enhance the ergonomy of search index

- Add missing elements (with same name) in the index

bug: 2356
Change-Id: I58b410f5870b34db7cfde7852e540511e1973afa
Signed-off-by: Faycal Abka <abka.faycal@gmail.com>

[2426] Improve detachment robustness

Bug: 2426

Change-Id: Ie3b3a363921546bfbe26a5f8389c4e5bd1455ee1
Signed-off-by: Thomas Guiu <thomas.guiu@soyatec.com>

[2419] [afm] disable cycle detection

Bug: 2419

Change-Id: I22a63fbde905f43753e2af4187e8b925b3dc1131
Signed-off-by: Thomas Guiu <thomas.guiu@soyatec.com>
(cherry picked from commit 5bac180eb917a2ace2a7026aded2ebcb685fc55c)

Signed-off-by: Arnaud Dieumegard <arnaud.dieumegard@obeo.fr>