#701 Use the new Sirius image dialog for the rich text editor
The image path returned by the selection component is relative to the
workspace and is serialized as it is.
To properly display the image in the embedded browser, the image path is
automatically converted into html with a path relative to the current
project.

MDEAddImageDialog is no longer used and has been deleted.

Change-Id: Ie9c11e19642f4d293ef0f42575a0f29bb3548d8c
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>