#753 Update the RichText editor status properly
The rich text editor status is now only based on the Session status.
In the scope of kitalpha/capella, AbstractMDERichTextWidget is only used
from MDERichTextEditor.isDirty().
So I consider that MDERichTextWidget.isDirty() was badly managed and can
be removed.
In the future, in the case we consider that MDERichTextWidget canc be
used independently of MDERichTextEditor, perhaps we could have
MDERichTextWidget.setDirty() and MDERichTextWidget.isDirty().
ListenerInstaller should
have BrowserFunction to listen to setText or insertText and call
MDERichTextWidget.setDirty(boolean).

https://github.com/eclipse/kitalpha/issues/753
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>