#753 Update the RichText editor status properly
The rich text editor status is now only based on the Session status.
In the scope of kitalpha/capella, AbstractMDERichTextWidget is only used
from MDERichTextEditor.isDirty().
So MDERichTextWidget.isDirty() can be removed.

Nevertheless, to keep the lock of the EObject being edited, a "change"
Listener has been added to be notified when the content of the editor is
changed.
The first change will call widget.saveContents() to be sure that the
model is modified a soon as possible. (before it was done a little
differently with the "key" listener)

Like eclipse "Save", now "Save All" properly reinitialize
setDirtyStateUpdated(false).

https://github.com/eclipse/kitalpha/issues/753
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>